requirements.txt
python-telegram-bot
requests
python-dotenv>=0.14.0